requirements.txt
tensorflow >=2.4.0
numpy
scipy
brian2
brian2hears
librosa
brain_pipe
torch
torchvision 
torchaudiopip